requirements.txt
black==24.3.0
opencv-python==4.10.0.84
keras
matplotlib==3.9.2
mypy==1.11.2
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
protobuf==3.20.3
pylint==3.2.7
ruff==0.6.4
scikit-learn==1.5.2
statsmodels==0.14.2
streamlit==1.38.0
torch --index-url https://download.pytorch.org/whl/cu124
torchvision --index-url https://download.pytorch.org/whl/cu124
torchaudio --index-url https://download.pytorch.org/whl/cu124
img2vec_pytorch==1.0.1
opencv-python==4.10.0.84
